more testing fixes
