+++ tools/psr/psrctl/cmd/list/list.go
@@ -60,7 +60,14 @@ func RunCmdList(cmd *cobra.Command, vzHelper helpers.VZHelper) error {
 
 	fmt.Println("Running Scenarios...")
 	for _, sc := range scenarios {
-		fmt.Println(sc.ID, sc.Description)
+		fmt.Println("----------------")
+		fmt.Printf("%s %s\n", "ID: ", sc.ID)
+		fmt.Printf("%s %s\n", "Description: ", sc.Description)
+		fmt.Println("Helm releases...")
+		for _, h := range sc.HelmReleases {
+			fmt.Printf("%s/%s\n", h.Namespace, h.Name)
+		}
+
 	}
 
 	return nil
